move contract features to base layer wallet
